[xDS e2e tests] move xDS bootstrap and resource helpers into their own library
