+++ test/cpp/end2end/xds/BUILD
@@ -46,6 +46,24 @@ grpc_cc_library(
 )
 
 grpc_cc_library(
+    name = "xds_utils",
+    testonly = True,
+    srcs = ["xds_utils.cc"],
+    hdrs = ["xds_utils.h"],
+    deps = [
+        ":xds_server",
+        "//src/proto/grpc/testing/xds/v3:cluster_proto",
+        "//src/proto/grpc/testing/xds/v3:discovery_proto",
+        "//src/proto/grpc/testing/xds/v3:endpoint_proto",
+        "//src/proto/grpc/testing/xds/v3:http_connection_manager_proto",
+        "//src/proto/grpc/testing/xds/v3:listener_proto",
+        "//src/proto/grpc/testing/xds/v3:route_proto",
+        "//src/proto/grpc/testing/xds/v3:router_proto",
+        "//test/core/util:grpc_test_util_base",
+    ],
+)
+
+grpc_cc_library(
     name = "xds_end2end_test_lib",
     testonly = True,
     srcs = ["xds_end2end_test_lib.cc"],
@@ -55,6 +73,7 @@ grpc_cc_library(
     ],
     deps = [
         ":xds_server",
+        ":xds_utils",
         "//:gpr",
         "//:grpc",
         "//:grpc++",
