+++ test/cpp/end2end/xds/xds_end2end_test_lib.h
@@ -50,6 +50,7 @@
 #include "test/cpp/end2end/counted_service.h"
 #include "test/cpp/end2end/test_service_impl.h"
 #include "test/cpp/end2end/xds/xds_server.h"
+#include "test/cpp/end2end/xds/xds_utils.h"
 
 namespace grpc {
 namespace testing {
@@ -201,29 +202,9 @@ class XdsTestType {
 // the indexes in the range [start_index, stop_index).  If stop_index
 // is 0, backends_.size() is used.  Backends may or may not be
 // xDS-enabled, at the discretion of the test.
-class XdsEnd2endTest : public ::testing::TestWithParam<XdsTestType> {
+class XdsEnd2endTest : public ::testing::TestWithParam<XdsTestType>,
+                       public XdsResourceUtils {
  protected:
-  using Cluster = ::envoy::config::cluster::v3::Cluster;
-  using ClusterLoadAssignment =
-      ::envoy::config::endpoint::v3::ClusterLoadAssignment;
-  using Listener = ::envoy::config::listener::v3::Listener;
-  using RouteConfiguration = ::envoy::config::route::v3::RouteConfiguration;
-  using HttpConnectionManager = ::envoy::extensions::filters::network::
-      http_connection_manager::v3::HttpConnectionManager;
-
-  // Default values for locality fields.
-  static const char kDefaultLocalityRegion[];
-  static const char kDefaultLocalityZone[];
-  static const uint32_t kDefaultLocalityWeight = 3;
-  static const int kDefaultLocalityPriority = 0;
-
-  // Default resource names.
-  static const char kServerName[];
-  static const char kDefaultRouteConfigurationName[];
-  static const char kDefaultClusterName[];
-  static const char kDefaultEdsServiceName[];
-  static const char kDefaultServerRouteConfigurationName[];
-
   // TLS certificate paths.
   static const char kCaCertPath[];
   static const char kServerCertPath[];
@@ -430,73 +411,6 @@ class XdsEnd2endTest : public ::testing::TestWithParam<XdsTestType> {
     std::shared_ptr<LrsServiceImpl> lrs_service_;
   };
 
-  // A builder for the xDS bootstrap config.
-  class BootstrapBuilder {
-   public:
-    BootstrapBuilder() {}
-    BootstrapBuilder& SetIgnoreResourceDeletion() {
-      ignore_resource_deletion_ = true;
-      return *this;
-    }
-    // If ignore_if_set is true, sets the default server only if it has
-    // not already been set.
-    BootstrapBuilder& SetDefaultServer(const std::string& server,
-                                       bool ignore_if_set = false) {
-      if (!ignore_if_set || top_server_.empty()) top_server_ = server;
-      return *this;
-    }
-    BootstrapBuilder& SetClientDefaultListenerResourceNameTemplate(
-        const std::string& client_default_listener_resource_name_template) {
-      client_default_listener_resource_name_template_ =
-          client_default_listener_resource_name_template;
-      return *this;
-    }
-    BootstrapBuilder& AddCertificateProviderPlugin(
-        const std::string& key, const std::string& name,
-        const std::string& plugin_config = "") {
-      plugins_[key] = {name, plugin_config};
-      return *this;
-    }
-    BootstrapBuilder& AddAuthority(
-        const std::string& authority, const std::string& server = "",
-        const std::string& client_listener_resource_name_template = "") {
-      authorities_[authority] = {server,
-                                 client_listener_resource_name_template};
-      return *this;
-    }
-    BootstrapBuilder& SetServerListenerResourceNameTemplate(
-        const std::string& server_listener_resource_name_template = "") {
-      server_listener_resource_name_template_ =
-          server_listener_resource_name_template;
-      return *this;
-    }
-
-    std::string Build();
-
-   private:
-    struct PluginInfo {
-      std::string name;
-      std::string plugin_config;
-    };
-    struct AuthorityInfo {
-      std::string server;
-      std::string client_listener_resource_name_template;
-    };
-
-    std::string MakeXdsServersText(absl::string_view server_uri);
-    std::string MakeNodeText();
-    std::string MakeCertificateProviderText();
-    std::string MakeAuthorityText();
-
-    bool ignore_resource_deletion_ = false;
-    std::string top_server_;
-    std::string client_default_listener_resource_name_template_;
-    std::map<std::string /*key*/, PluginInfo> plugins_;
-    std::map<std::string /*authority_name*/, AuthorityInfo> authorities_;
-    std::string server_listener_resource_name_template_ =
-        "grpc/server?xds.resource.listening_address=%s";
-  };
-
   // RPC services used to talk to the backends.
   enum RpcService {
     SERVICE_ECHO,
@@ -525,47 +439,17 @@ class XdsEnd2endTest : public ::testing::TestWithParam<XdsTestType> {
   // balancer_, which is already populated with default resources.
   std::unique_ptr<BalancerServerThread> CreateAndStartBalancer();
 
-  // Returns the name of the server-side xDS Listener resource for a
-  // backend on the specified port.
-  std::string GetServerListenerName(int port);
-
-  // Returns a copy of listener_template with the server-side resource
-  // name and the port in the socket address populated.
-  Listener PopulateServerListenerNameAndPort(const Listener& listener_template,
-                                             int port);
-
-  // Interface for accessing HttpConnectionManager config in Listener.
-  class HcmAccessor {
-   public:
-    virtual ~HcmAccessor() = default;
-    virtual HttpConnectionManager Unpack(const Listener& listener) const = 0;
-    virtual void Pack(const HttpConnectionManager& hcm,
-                      Listener* listener) const = 0;
-  };
-
-  // Client-side impl.
-  class ClientHcmAccessor : public HcmAccessor {
-   public:
-    HttpConnectionManager Unpack(const Listener& listener) const override;
-    void Pack(const HttpConnectionManager& hcm,
-              Listener* listener) const override;
-  };
-
-  // Server-side impl.
-  class ServerHcmAccessor : public HcmAccessor {
-   public:
-    HttpConnectionManager Unpack(const Listener& listener) const override;
-    void Pack(const HttpConnectionManager& hcm,
-              Listener* listener) const override;
-  };
-
   // Sets the Listener and RouteConfiguration resource on the specified
   // balancer.  If RDS is in use, they will be set as separate resources;
   // otherwise, the RouteConfig will be inlined into the Listener.
   void SetListenerAndRouteConfiguration(
       BalancerServerThread* balancer, Listener listener,
       const RouteConfiguration& route_config,
-      const HcmAccessor& hcm_accessor = ClientHcmAccessor());
+      const HcmAccessor& hcm_accessor = ClientHcmAccessor()) {
+    XdsResourceUtils::SetListenerAndRouteConfiguration(
+        balancer->ads_service(), std::move(listener), route_config,
+        GetParam().enable_rds_testing(), hcm_accessor);
+  }
 
   // A convenient wrapper for setting the Listener and
   // RouteConfiguration resources on the server side.
@@ -583,53 +467,11 @@ class XdsEnd2endTest : public ::testing::TestWithParam<XdsTestType> {
   // (either listener_to_copy, or if that is null, default_listener_).
   void SetRouteConfiguration(BalancerServerThread* balancer,
                              const RouteConfiguration& route_config,
-                             const Listener* listener_to_copy = nullptr);
-
-  // Arguments for constructing an EDS resource.
-  struct EdsResourceArgs {
-    // An individual endpoint for a backend running on a specified port.
-    struct Endpoint {
-      explicit Endpoint(int port,
-                        ::envoy::config::core::v3::HealthStatus health_status =
-                            ::envoy::config::core::v3::HealthStatus::UNKNOWN,
-                        int lb_weight = 1,
-                        std::vector<int> additional_ports = {})
-          : port(port),
-            health_status(health_status),
-            lb_weight(lb_weight),
-            additional_ports(std::move(additional_ports)) {}
-
-      int port;
-      ::envoy::config::core::v3::HealthStatus health_status;
-      int lb_weight;
-      std::vector<int> additional_ports;
-    };
-
-    // A locality.
-    struct Locality {
-      Locality(std::string sub_zone, std::vector<Endpoint> endpoints,
-               uint32_t lb_weight = kDefaultLocalityWeight,
-               int priority = kDefaultLocalityPriority)
-          : sub_zone(std::move(sub_zone)),
-            endpoints(std::move(endpoints)),
-            lb_weight(lb_weight),
-            priority(priority) {}
-
-      const std::string sub_zone;
-      std::vector<Endpoint> endpoints;
-      uint32_t lb_weight;
-      int priority;
-    };
-
-    EdsResourceArgs() = default;
-    explicit EdsResourceArgs(std::vector<Locality> locality_list)
-        : locality_list(std::move(locality_list)) {}
-
-    std::vector<Locality> locality_list;
-    std::map<std::string, uint32_t> drop_categories;
-    ::envoy::type::v3::FractionalPercent::DenominatorType drop_denominator =
-        ::envoy::type::v3::FractionalPercent::MILLION;
-  };
+                             const Listener* listener_to_copy = nullptr) {
+    XdsResourceUtils::SetRouteConfiguration(balancer->ads_service(),
+                                    route_config, GetParam().enable_rds_testing(),
+                                    listener_to_copy);
+  }
 
   // Helper method for generating an endpoint for a backend, for use in
   // constructing an EDS resource.
@@ -661,11 +503,6 @@ class XdsEnd2endTest : public ::testing::TestWithParam<XdsTestType> {
   EdsResourceArgs::Endpoint MakeNonExistantEndpoint() {
     return EdsResourceArgs::Endpoint(grpc_pick_unused_port_or_die());
   }
-
-  // Constructs an EDS resource.
-  ClusterLoadAssignment BuildEdsResource(
-      const EdsResourceArgs& args,
-      absl::string_view eds_service_name = kDefaultEdsServiceName);
 
   //
   // Backend management
@@ -727,7 +564,7 @@ class XdsEnd2endTest : public ::testing::TestWithParam<XdsTestType> {
   // Initializes global state for the client, such as the bootstrap file
   // and channel args for the XdsClient.  Then calls ResetStub().
   // All tests must call this exactly once at the start of the test.
-  void InitClient(BootstrapBuilder builder = BootstrapBuilder(),
+  void InitClient(XdsBootstrapBuilder builder = XdsBootstrapBuilder(),
                   std::string lb_expected_authority = "",
                   int xds_resource_does_not_exist_timeout_ms = 0);
